remove the removing routine when ffmpeg fails

+++ subfuncs/youtube_explorer.py
@@ -174,8 +174,9 @@ def convertMP3(baseDIR, old_filename, new_filename):
   mp3_path = os.path.join(baseDIR, new_filename)
   old_path = os.path.join(baseDIR, old_filename)
   if not os.path.exists(mp3_path):
-    subprocess.call(['ffmpeg', '-i', old_path, mp3_path])
-    os.remove(old_path)
+    ret = subprocess.call(['ffmpeg', '-i', old_path, mp3_path])
+    if ret == 0:
+      os.remove(old_path)
 
 def getSongFromYouTube(artist, title, songID, lyric, albumID, baseMusicDir, baseImageDir,
                        isOverwriteMode=False, music_reporter=None):
